+++ Geode/Errors/ForkedElseError.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Geode.Errors
+{
+    public class ForkedElseError() : GeodeError("Else statements for branching execute statements is not supported yet")
+    {
+        
+    }
+}