+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* 
+* Fixed an issue with trying to get a property from a macro value.
+* Fixed an issue where score operations weren't always correctly stored as scores, causing an error.
+* Fixed an issue with forking `execute` statements not actually forking.